+++ src/main/kotlin/com/coder/gateway/cli/CoderCLIManager.kt
@@ -1,41 +1,42 @@
 package com.coder.gateway.cli
 
+import com.coder.gateway.cli.downloader.CoderDownloadApi
+import com.coder.gateway.cli.downloader.CoderDownloadService
+import com.coder.gateway.cli.downloader.DownloadResult
 import com.coder.gateway.cli.ex.MissingVersionException
-import com.coder.gateway.cli.ex.ResponseException
 import com.coder.gateway.cli.ex.SSHConfigFormatException
+import com.coder.gateway.cli.ex.UnsignedBinaryExecutionDeniedException
+import com.coder.gateway.cli.gpg.GPGVerifier
+import com.coder.gateway.cli.gpg.VerificationResult
 import com.coder.gateway.sdk.v2.models.User
 import com.coder.gateway.sdk.v2.models.Workspace
 import com.coder.gateway.sdk.v2.models.WorkspaceAgent
 import com.coder.gateway.settings.CoderSettings
 import com.coder.gateway.settings.CoderSettingsState
 import com.coder.gateway.util.CoderHostnameVerifier
+import com.coder.gateway.util.DialogUi
 import com.coder.gateway.util.InvalidVersionException
-import com.coder.gateway.util.OS
 import com.coder.gateway.util.SemVer
 import com.coder.gateway.util.coderSocketFactory
+import com.coder.gateway.util.coderTrustManagers
 import com.coder.gateway.util.escape
 import com.coder.gateway.util.escapeSubcommand
-import com.coder.gateway.util.getHeaders
-import com.coder.gateway.util.getOS
 import com.coder.gateway.util.safeHost
-import com.coder.gateway.util.sha1
 import com.intellij.openapi.diagnostic.Logger
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import com.squareup.moshi.JsonDataException
 import com.squareup.moshi.Moshi
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import okhttp3.OkHttpClient
 import org.zeroturnaround.exec.ProcessExecutor
+import retrofit2.Retrofit
 import java.io.EOFException
-import java.io.FileInputStream
 import java.io.FileNotFoundException
-import java.net.ConnectException
-import java.net.HttpURLConnection
 import java.net.URL
-import java.nio.file.Files
 import java.nio.file.Path
-import java.nio.file.StandardCopyOption
-import java.util.zip.GZIPInputStream
-import javax.net.ssl.HttpsURLConnection
+import javax.net.ssl.X509TrustManager
 
 /**
  * Version output from the CLI's version command.
@@ -57,7 +58,7 @@ internal data class Version(
  * 6. Since the binary directory can be read-only, if downloading fails, start
  *    from step 2 with the data directory.
  */
-fun ensureCLI(
+suspend fun ensureCLI(
     deploymentURL: URL,
     buildVersion: String,
     settings: CoderSettings,
@@ -72,6 +73,7 @@ fun ensureCLI(
     // the 304 method.
     val cliMatches = cli.matchesVersion(buildVersion)
     if (cliMatches == true) {
+        indicator?.invoke("Local CLI version matches server version: $buildVersion")
         return cli
     }
 
@@ -79,7 +81,7 @@ fun ensureCLI(
     if (settings.enableDownloads) {
         indicator?.invoke("Downloading Coder CLI...")
         try {
-            cli.download()
+            cli.download(buildVersion, indicator)
             return cli
         } catch (e: java.nio.file.AccessDeniedException) {
             // Might be able to fall back to the data directory.
@@ -95,12 +97,13 @@ fun ensureCLI(
     val dataCLI = CoderCLIManager(deploymentURL, settings, true)
     val dataCLIMatches = dataCLI.matchesVersion(buildVersion)
     if (dataCLIMatches == true) {
+        indicator?.invoke("Local CLI version from data directory matches server version: $buildVersion")
         return dataCLI
     }
 
     if (settings.enableDownloads) {
-        indicator?.invoke("Downloading Coder CLI...")
-        dataCLI.download()
+        indicator?.invoke("Downloading Coder CLI to the data directory...")
+        dataCLI.download(buildVersion, indicator)
         return dataCLI
     }
 
@@ -128,78 +131,147 @@ class CoderCLIManager(
     private val settings: CoderSettings = CoderSettings(CoderSettingsState()),
     // If the binary directory is not writable, this can be used to force the
     // manager to download to the data directory instead.
-    forceDownloadToData: Boolean = false,
+    private val forceDownloadToData: Boolean = false,
 ) {
+    private val downloader = createDownloadService()
+    private val gpgVerifier = GPGVerifier(settings)
+
     val remoteBinaryURL: URL = settings.binSource(deploymentURL)
     val localBinaryPath: Path = settings.binPath(deploymentURL, forceDownloadToData)
     val coderConfigPath: Path = settings.dataDir(deploymentURL).resolve("config")
 
-    /**
-     * Download the CLI from the deployment if necessary.
-     */
-    fun download(): Boolean {
-        val eTag = getBinaryETag()
-        val conn = remoteBinaryURL.openConnection() as HttpURLConnection
-        if (settings.headerCommand.isNotBlank()) {
-            val headersFromHeaderCommand = getHeaders(deploymentURL, settings.headerCommand)
-            for ((key, value) in headersFromHeaderCommand) {
-                conn.setRequestProperty(key, value)
-            }
-        }
-        if (eTag != null) {
-            logger.info("Found existing binary at $localBinaryPath; calculated hash as $eTag")
-            conn.setRequestProperty("If-None-Match", "\"$eTag\"")
-        }
-        conn.setRequestProperty("Accept-Encoding", "gzip")
-        if (conn is HttpsURLConnection) {
-            conn.sslSocketFactory = coderSocketFactory(settings.tls)
-            conn.hostnameVerifier = CoderHostnameVerifier(settings.tls.altHostname)
-        }
+    private fun createDownloadService(): CoderDownloadService {
+        val okHttpClient = OkHttpClient.Builder()
+            .sslSocketFactory(
+                coderSocketFactory(settings.tls),
+                coderTrustManagers(settings.tls.caPath)[0] as X509TrustManager
+            )
+            .hostnameVerifier(CoderHostnameVerifier(settings.tls.altHostname))
+            .build()
 
-        try {
-            conn.connect()
-            logger.info("GET ${conn.responseCode} $remoteBinaryURL")
-            when (conn.responseCode) {
-                HttpURLConnection.HTTP_OK -> {
-                    logger.info("Downloading binary to $localBinaryPath")
-                    Files.createDirectories(localBinaryPath.parent)
-                    conn.inputStream.use {
-                        Files.copy(
-                            if (conn.contentEncoding == "gzip") GZIPInputStream(it) else it,
-                            localBinaryPath,
-                            StandardCopyOption.REPLACE_EXISTING,
-                        )
-                    }
-                    if (getOS() != OS.WINDOWS) {
-                        localBinaryPath.toFile().setExecutable(true)
-                    }
-                    return true
-                }
+        val retrofit = Retrofit.Builder()
+            .baseUrl(deploymentURL.toString())
+            .client(okHttpClient)
+            .build()
 
-                HttpURLConnection.HTTP_NOT_MODIFIED -> {
-                    logger.info("Using cached binary at $localBinaryPath")
-                    return false
-                }
-            }
-        } catch (e: ConnectException) {
-            // Add the URL so this is more easily debugged.
-            throw ConnectException("${e.message} to $remoteBinaryURL")
-        } finally {
-            conn.disconnect()
-        }
-        throw ResponseException("Unexpected response from $remoteBinaryURL", conn.responseCode)
+        val service = retrofit.create(CoderDownloadApi::class.java)
+        return CoderDownloadService(settings, service, deploymentURL, forceDownloadToData)
     }
 
     /**
-     * Return the entity tag for the binary on disk, if any.
+     * Download the CLI from the deployment if necessary.
      */
-    private fun getBinaryETag(): String? = try {
-        sha1(FileInputStream(localBinaryPath.toFile()))
-    } catch (e: FileNotFoundException) {
-        null
-    } catch (e: Exception) {
-        logger.warn("Unable to calculate hash for $localBinaryPath", e)
-        null
+    suspend fun download(buildVersion: String, showTextProgress: ((t: String) -> Unit)? = null): Boolean {
+        try {
+            val cliResult = withContext(Dispatchers.IO) {
+                downloader.downloadCli(buildVersion, showTextProgress)
+            }.let { result ->
+                when {
+                    result.isSkipped() -> return false
+                    result.isNotFound() -> throw IllegalStateException("Could not find Coder CLI")
+                    result.isFailed() -> throw (result as DownloadResult.Failed).error
+                    else -> result as DownloadResult.Downloaded
+                }
+            }
+
+            var signatureResult = withContext(Dispatchers.IO) {
+                downloader.downloadSignature(showTextProgress)
+            }
+
+            if (signatureResult.isNotDownloaded()) {
+                if (settings.fallbackOnCoderForSignatures) {
+                    logger.info("Trying to download signature file from releases.coder.com")
+                    signatureResult = withContext(Dispatchers.IO) {
+                        downloader.downloadReleasesSignature(buildVersion, showTextProgress)
+                    }
+
+                    // if we could still not download it, ask the user if he accepts the risk
+                    if (signatureResult.isNotDownloaded()) {
+                        val acceptsUnsignedBinary = DialogUi(settings)
+                            .confirm(
+                                "Security Warning",
+                                "Could not fetch any signatures for ${cliResult.source} from releases.coder.com. Would you like to run it anyway?"
+                            )
+
+                        if (acceptsUnsignedBinary) {
+                            downloader.commit()
+                            return true
+                        } else {
+                            throw UnsignedBinaryExecutionDeniedException("Running unsigned CLI from ${cliResult.source} was denied by the user")
+                        }
+                    }
+                } else {
+                    // we are not allowed to fetch signatures from releases.coder.com
+                    // so we will ask the user if he wants to continue
+                    val acceptsUnsignedBinary = DialogUi(settings)
+                        .confirm(
+                            "Security Warning",
+                            "No signatures were found for ${cliResult.source} and fallback to releases.coder.com is not allowed. Would you like to run it anyway?"
+                        )
+
+                    if (acceptsUnsignedBinary) {
+                        downloader.commit()
+                        return true
+                    } else {
+                        throw UnsignedBinaryExecutionDeniedException("Running unsigned CLI from ${cliResult.source} was denied by the user")
+                    }
+                }
+            }
+
+            // we have the cli, and signature is downloaded, let's verify the signature
+            signatureResult = signatureResult as DownloadResult.Downloaded
+            gpgVerifier.verifySignature(cliResult.dst, signatureResult.dst).let { result ->
+                when {
+                    result.isValid() -> {
+                        downloader.commit()
+                        return true
+                    }
+
+                    else -> {
+                        logFailure(result, cliResult, signatureResult)
+                        // prompt the user if he wants to accept the risk
+                        val shouldRunAnyway = DialogUi(settings)
+                            .confirm(
+                                "Security Warning",
+                                "Could not verify the authenticity of the ${cliResult.source}, it may be tampered with. Would you like to run it anyway?"
+                            )
+
+                        if (shouldRunAnyway) {
+                            downloader.commit()
+                            return true
+                        } else {
+                            throw UnsignedBinaryExecutionDeniedException("Running unverified CLI from ${cliResult.source} was denied by the user")
+                        }
+                    }
+                }
+            }
+        } finally {
+            downloader.cleanup()
+        }
+    }
+
+    private fun logFailure(
+        result: VerificationResult,
+        cliResult: DownloadResult.Downloaded,
+        signatureResult: DownloadResult.Downloaded
+    ) {
+        when {
+            result.isInvalid() -> {
+                val reason = (result as VerificationResult.Invalid).reason
+                logger.error("Signature of ${cliResult.dst} is invalid." + reason?.let { " Reason: $it" }
+                    .orEmpty())
+            }
+
+            result.signatureIsNotFound() -> {
+                logger.error("Can't verify signature of ${cliResult.dst} because ${signatureResult.dst} does not exist")
+            }
+
+            else -> {
+                UnsignedBinaryExecutionDeniedException((result as DownloadResult.Failed).error.message)
+                val failure = result as DownloadResult.Failed
+                logger.error("Failed to verify signature for ${cliResult.dst}", failure.error)
+            }
+        }
     }
 
     /**
@@ -278,7 +350,8 @@ class CoderCLIManager(
             if (settings.sshLogDirectory.isNotBlank()) escape(settings.sshLogDirectory) else null,
             if (feats.reportWorkspaceUsage) "--usage-app=jetbrains" else null,
         )
-        val backgroundProxyArgs = baseArgs + listOfNotNull(if (feats.reportWorkspaceUsage) "--usage-app=disable" else null)
+        val backgroundProxyArgs =
+            baseArgs + listOfNotNull(if (feats.reportWorkspaceUsage) "--usage-app=disable" else null)
         val extraConfig =
             if (settings.sshConfigOptions.isNotBlank()) {
                 "\n" + settings.sshConfigOptions.prependIndent("  ")
@@ -295,22 +368,22 @@ class CoderCLIManager(
         val blockContent =
             if (feats.wildcardSSH) {
                 startBlock + System.lineSeparator() +
-                    """
+                        """
                     Host ${getHostPrefix()}--*
                       ProxyCommand ${proxyArgs.joinToString(" ")} --ssh-host-prefix ${getHostPrefix()}-- %h
                     """.trimIndent()
-                        .plus("\n" + sshOpts.prependIndent("  "))
-                        .plus(extraConfig)
-                        .plus("\n\n")
-                        .plus(
-                            """
+                            .plus("\n" + sshOpts.prependIndent("  "))
+                            .plus(extraConfig)
+                            .plus("\n\n")
+                            .plus(
+                                """
                             Host ${getHostPrefix()}-bg--*
                               ProxyCommand ${backgroundProxyArgs.joinToString(" ")} --ssh-host-prefix ${getHostPrefix()}-bg-- %h
                             """.trimIndent()
-                                .plus("\n" + sshOpts.prependIndent("  "))
-                                .plus(extraConfig),
-                        ).replace("\n", System.lineSeparator()) +
-                    System.lineSeparator() + endBlock
+                                    .plus("\n" + sshOpts.prependIndent("  "))
+                                    .plus(extraConfig),
+                            ).replace("\n", System.lineSeparator()) +
+                        System.lineSeparator() + endBlock
             } else if (workspaceNames.isEmpty()) {
                 ""
             } else {
@@ -329,7 +402,12 @@ class CoderCLIManager(
                             .plus(
                                 """
                             Host ${getBackgroundHostName(it.first, currentUser, it.second)}
-                              ProxyCommand ${backgroundProxyArgs.joinToString(" ")} ${getWorkspaceParts(it.first, it.second)}
+                              ProxyCommand ${backgroundProxyArgs.joinToString(" ")} ${
+                                    getWorkspaceParts(
+                                        it.first,
+                                        it.second
+                                    )
+                                }
                                 """.trimIndent()
                                     .plus("\n" + sshOpts.prependIndent("  "))
                                     .plus(extraConfig),
@@ -443,6 +521,7 @@ class CoderCLIManager(
             is InvalidVersionException -> {
                 logger.info("Got invalid version from $localBinaryPath: ${e.message}")
             }
+
             else -> {
                 // An error here most likely means the CLI does not exist or
                 // it executed successfully but output no version which