+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ## Unreleased
 
+### Added
+
+- support for checking if CLI is signed
+- improved progress reporting while downloading the CLI
+
 ## 2.21.1 - 2025-06-26
 
 ### Fixed